# fhouritu_soudann/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for legal consultation booking
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BookingSlot {
  date: string;
  /** Number of available legal consultation slots */
  legalSlots: number;
  /** Number of booked legal consultation slots */
  legalBooked: number;
}

export interface Booking {
  id: number;
  date: string;
  time: string;
  name: string;
  email: string;
  phone?: string;
  notes?: string;
  createdAt: string;
}

export interface CreateBookingRequest {
  date: string;
  time: string;
  name: string;
  email: string;
  phone?: string;
  notes?: string;
}

export interface TimeSlot {
  time: string;
  available: boolean;
  booked: boolean;
}

export interface DateAvailability {
  date: string;
  slots: TimeSlot[];
}

export interface ErrorResponse {
  error: string;
}

export type GetBookingsParams = {
  year?: number;
  month?: number;
};

export type GetAvailabilityParams = {
  date: string;
};